refactor(api): route HTTP endpoints through LocalBackend
- Delete duplicate getDbConfigFromEntry (canonical getDbConfig in repo-manager)
- Delete buildGraph function (moved to LocalBackend.getGraphForApi)
- /api/query routes through backend.executeCypher()
- /api/search routes through new backend.searchForApi()
- /api/graph routes through new backend.getGraphForApi()
- Returns 501 for graph viz on Neptune repos (proper error)
- Eliminates per-request NeptuneAdapter construction
- Remove double closeLbug() in shutdown handler
- Update Neptune injection safety test to scan new code location
